Reduce chance of state getting stale while reading from phone
Reading takes about 2s, so when this action is called multiple times
(because something was broken) every separate call that reads the
state within those 2s will read a change and add it as new state.

This change reduces the chance of the collision, because reading and
writing the updated state should never take 2s. However, applying
a RepeatableRead transaction around reading the state and writing
new state will prevent the problem altogether.